refactor: optimize file structure and exception implementation
