+++ src/parser/singles.cpp
@@ -1,14 +1,53 @@
 /**
  * @file src/parser/singles.cpp
  * @brief Parser implementation for single tokens.
- * @author The-XiaoBai
- * @date 2026/01/31
 **/
 
 #include "parser.hpp"
 #include "utils.hpp"
 
 namespace DemoLang {
+
+/**
+ * @brief Flyweight factory for AST nodes
+ */
+class ASTFlyweight {
+private:
+    static Utils::FlyweightFactory<std::string, ASTNode>& factory() {
+        return Utils::FlyweightFactory<std::string, ASTNode>::instance();
+    }
+    
+public:
+    static std::shared_ptr<ASTNode> getIdNode(const std::string& name) {
+        return factory().getFlyweight("id:" + name, [&name]() {
+            return std::make_shared<IdNode>(name);
+        });
+    }
+    
+    static std::shared_ptr<ASTNode> getIntNode(long long value) {
+        std::string key = "int:" + std::to_string(value);
+        return factory().getFlyweight(key, [value]() {
+            return std::make_shared<IntNode>(value);
+        });
+    }
+    
+    static std::shared_ptr<ASTNode> getFloatNode(long double value) {
+        std::string key = "float:" + std::to_string(value);
+        return factory().getFlyweight(key, [value]() {
+            return std::make_shared<FloatNode>(value);
+        });
+    }
+    
+    static std::shared_ptr<ASTNode> getStringNode(const std::string& value) {
+        std::string key = "string:" + value;
+        return factory().getFlyweight(key, [&value]() {
+            return std::make_shared<StringNode>(value);
+        });
+    }
+    
+    static void clearCache() { factory().clear(); }
+    static size_t cacheSize() { return factory().size(); }
+};
 
 // AST Node Factory using Utils Factory template
 class ASTNodeFactory : public Utils::Factory<TokenType, ASTNode>, public Utils::Singleton<ASTNodeFactory> {
@@ -45,6 +84,8 @@ public:
                     return createParenthesizedNode(token, parser);
                 }
                 return std::make_shared<ErrorNode>("Unexpected operator: " + token.value);
+            case TokenType::ERROR:
+                return createErrorNode(token.value);
             default:
                 return std::make_shared<ErrorNode>("Unexpected token: " + token.value);
         }
